Refactor home feature: Remove unused widgets and implement voice logging service
- Deleted MealsSection, MicButtonWidget, RecordingTimerWidget, VoiceRecordingFab, and related files to streamline the codebase.
- Introduced VoiceLoggingService to manage voice recording functionality.
- Updated GoalPlanModel to use consistent ID naming conventions.
- Added CaloriesCard, MacroItem, and MacronutrientsCard for improved calorie and macro tracking UI.
- Implemented HomeScreenBlocListener and HomeScreenContent for better state management with Bloc.
- Created LoggedMealsList and LoggedMealCard for displaying logged meals.
- Added EmptyMealsState for handling empty meal logging scenarios.
