--- lib/features/home/data/models/nutrition_data.dart
@@ -1,21 +1,0 @@
-class NutritionData {
-  final double calories;
-  final double protein;
-  final double carbs;
-  final double fat;
-  final double caloriesGoal;
-  final double proteinGoal;
-  final double carbsGoal;
-  final double fatGoal;
-
-  const NutritionData({
-    required this.calories,
-    required this.protein,
-    required this.carbs,
-    required this.fat,
-    required this.caloriesGoal,
-    required this.proteinGoal,
-    required this.carbsGoal,
-    required this.fatGoal,
-  });
-}
