+++ lib/features/home/presentation/widgets/calories_and_macros/calories_card.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../../../core/theme/app_colors.dart';
-import '../../../../core/theme/app_text_styles.dart';
-import '../../../../core/theme/app_design_system.dart';
-import '../../../../core/helpers/spacing.dart';
+import '../../../../../core/theme/app_colors.dart';
+import '../../../../../core/theme/app_text_styles.dart';
+import '../../../../../core/theme/app_design_system.dart';
+import '../../../../../core/helpers/spacing.dart';
 
 class CaloriesCard extends StatelessWidget {
   final int consumed;
