+++ lib/features/home/presentation/widgets/logged_meals/empty_meals_state.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../../../core/helpers/spacing.dart';
+import '../../../../../core/helpers/spacing.dart';
 
 class EmptyMealsState extends StatelessWidget {
   const EmptyMealsState({super.key});
