+++ lib/features/home/presentation/widgets/logged_meals/logged_meal_card.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../../../core/helpers/spacing.dart';
-import '../../../../core/theme/app_colors.dart';
-import '../../../../core/theme/app_text_styles.dart';
-import '../../data/models/meal_data.dart';
+import '../../../../../core/helpers/spacing.dart';
+import '../../../../../core/theme/app_colors.dart';
+import '../../../../../core/theme/app_text_styles.dart';
+import '../../../data/models/meal_data.dart';
 
 class LoggedMealCard extends StatelessWidget {
   final MealData meal;
