+++ lib/features/home/presentation/widgets/logged_meals/meals_section.dart
@@ -1,17 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import '../../../../core/theme/app_text_styles.dart';
-import '../../../../core/theme/app_colors.dart';
-import '../../../../core/helpers/spacing.dart';
+import '../../../../../core/theme/app_text_styles.dart';
+import '../../../../../core/theme/app_colors.dart';
+import '../../../../../core/helpers/spacing.dart';
 import 'meal_card.dart';
 
 class MealsSection extends StatelessWidget {
   final List<Map<String, dynamic>> meals;
 
-  const MealsSection({
-    super.key,
-    required this.meals,
-  });
+  const MealsSection({super.key, required this.meals});
 
   Widget _buildMealsList() {
     if (meals.isEmpty) {
